Writing the webscraping code
This is a later stage addition of the web scraping code I have written for the project. The code was written and tested first in a Juptyer notebook before I copy and pasted it into a .py file so that it could be executed more easily. More changes to follow